More acceptable but still has problems

+++ commonvars.h
@@ -1,3 +1,3 @@
-extern float IDENTITY[16];
+extern double IDENTITY[16];
 extern matrix mainMatrix;
 extern matrix matrixStack[32];
